feat(chat): move model controls into the composer
