+++ crates/agent-gui/test/chat/execution-mode-model-picker.test.mjs
@@ -2,15 +2,23 @@ import assert from "node:assert/strict";
 import { readFileSync } from "node:fs";
 import test from "node:test";
 
-const headerSources = [
+const pickerSources = [
   readFileSync(
-    new URL("../../../agent-ui/src/components/chat/ChatHeader.tsx", import.meta.url),
+    new URL("../../../agent-ui/src/components/chat/ComposerModelControls.tsx", import.meta.url),
     "utf8",
   ),
 ];
+const headerSource = readFileSync(
+  new URL("../../../agent-ui/src/components/chat/ChatHeader.tsx", import.meta.url),
+  "utf8",
+);
+const composerSource = readFileSync(
+  new URL("../../../agent-ui/src/pages/chat/ChatComposerBar.tsx", import.meta.url),
+  "utf8",
+);
 
 test("model pickers use popover semantics instead of menu semantics", () => {
-  for (const source of headerSources) {
+  for (const source of pickerSources) {
     assert.match(source, /import \{ Popover \} from "@base-ui\/react"/);
     assert.match(source, /<Popover\.Root open=\{isModelPickerOpen\}/);
     assert.match(source, /<Popover\.Popup/);
@@ -20,7 +28,7 @@ test("model pickers use popover semantics instead of menu semantics", () => {
 });
 
 test("execution mode switchers expose a native radio group", () => {
-  for (const source of headerSources) {
+  for (const source of pickerSources) {
     assert.match(source, /role="radiogroup"/);
     assert.match(source, /aria-label=\{t\("settings\.executionMode"\)\}/);
     assert.equal((source.match(/type="radio"/g) ?? []).length, 2);
@@ -33,7 +41,7 @@ test("execution mode switchers expose a native radio group", () => {
 });
 
 test("popover interactions preserve mode changes and close after model selection", () => {
-  for (const source of headerSources) {
+  for (const source of pickerSources) {
     assert.match(source, /onClick=\{\(\) => toggleGroup\(group\.id\)\}/);
     assert.match(source, /aria-pressed=\{isSelected\}/);
     assert.match(source, /onSelectModel\(parsed\);\s+setIsModelPickerOpen\(false\);/);
@@ -41,7 +49,7 @@ test("popover interactions preserve mode changes and close after model selection
 });
 
 test("model pickers search models and providers", () => {
-  for (const source of headerSources) {
+  for (const source of pickerSources) {
     assert.match(source, /initialFocus=\{searchInputRef\}/);
     assert.match(source, /placeholder=\{t\("chat\.searchModel"\)\}/);
     assert.match(source, /\w+\.model\.toLowerCase\(\)\.includes\(normalizedSearch\)/);
@@ -51,7 +59,7 @@ test("model pickers search models and providers", () => {
 });
 
 test("provider groups reveal the edit affordance before the count on hover", () => {
-  for (const source of headerSources) {
+  for (const source of pickerSources) {
     assert.match(source, /\bPencil\b/);
     assert.match(source, /t\("settings\.editProvider"\)/);
     assert.doesNotMatch(source, /title=\{`\$\{t\("settings\.editProvider"\)/);
@@ -64,5 +72,22 @@ test("provider groups reveal the edit affordance before the count on hover", () 
       source,
       /setIsModelPickerOpen\(false\);\s+onOpenSettings\("providers", group\.id\);/,
     );
+  }
+});
+
+test("model and runtime controls are colocated in the composer instead of the header", () => {
+  assert.match(composerSource, /<ComposerModelControls/);
+  assert.match(composerSource, /<RuntimeControlTooltip label=\{uploadTooltip\}>/);
+  assert.ok(
+    composerSource.indexOf("<ComposerModelControls") <
+      composerSource.indexOf("<RuntimeControlTooltip label={uploadTooltip}>"),
+  );
+  assert.doesNotMatch(headerSource, /model-selector-trigger|Popover\.Root|currentModelLabel/);
+
+  for (const source of pickerSources) {
+    assert.match(source, /aria-label=\{t\("chat\.runtime\.controls"\)\}/);
+    assert.match(source, /nativeWebSearchEnabled: !chatRuntimeControls\.nativeWebSearchEnabled/);
+    assert.match(source, /thinkingEnabled: !chatRuntimeControls\.thinkingEnabled/);
+    assert.match(source, /onChatRuntimeControlsChange\(\{ reasoning:/);
   }
 });
