+++ crates/agent-ui/src/i18n/translations/enUSCommon.ts
@@ -256,6 +256,9 @@ export const EN_US_COMMON_TRANSLATIONS = {
   "chat.runtime.webSearchOn": "Web search enabled",
   "chat.runtime.webSearchOff": "Web search disabled",
   "chat.runtime.webSearchTooltip": "Toggle web search",
+  "chat.runtime.controls": "Model controls",
+  "chat.runtime.webSearch": "Web",
+  "chat.runtime.thinking": "Thinking",
   "chat.runtime.tunnelToolAvailable": "Tunnel tool available",
   "chat.runtime.tunnelToolUnavailable": "Tunnel tool unavailable",
   "chat.runtime.tunnelAgentModeRequired": "Tunnel tool requires Agent mode",
