+++ crates/agent-ui/src/i18n/translations/zhCNCommon.ts
@@ -231,6 +231,9 @@ export const ZH_CN_COMMON_TRANSLATIONS = {
   "chat.runtime.webSearchOn": "联网搜索已开启",
   "chat.runtime.webSearchOff": "联网搜索已关闭",
   "chat.runtime.webSearchTooltip": "联网搜索",
+  "chat.runtime.controls": "模型与能力",
+  "chat.runtime.webSearch": "联网",
+  "chat.runtime.thinking": "思考",
   "chat.runtime.tunnelToolAvailable": "内网穿透工具已可用",
   "chat.runtime.tunnelToolUnavailable": "内网穿透工具不可用",
   "chat.runtime.tunnelAgentModeRequired": "Agent 模式下可使用内网穿透工具",
